==> notebooks/01_data_exploration.ipynb <==
import pandas as pd
from geopy.distance import geodesic







nb_df = pd.read_csv('../data/raw/nb.csv')
location_df = pd.read_csv('../data/raw/location.csv')

nb_df.info()
nb_df.describe()
nb_df.head(5)



nb_df.info()
nb_df.describe()
location_df.head(5)

all_nan_columns_nb = nb_df.columns[nb_df.isna().all()].tolist()
all_nan_columns_location = location_df.columns[location_df.isna().all()].tolist()

if all_nan_columns_nb:
    print("\nColumns in nb_df with all NaN values:")
    for col in all_nan_columns_nb:
        print(col)
else:
    print("\nNo columns in nb_df have all NaN values.")

if all_nan_columns_location:
    print("\nColumns in location_df with all NaN values:")
    for col in all_nan_columns_location:
        print(col)
else:
    print("\nNo columns in location_df have all NaN values.")


nb_df_cleaned = nb_df.dropna(axis=1, how='all')
location_df_cleaned = location_df.dropna(axis=1, how='all')

nb_df_cleaned = nb_df_cleaned.dropna()
location_df_cleaned = location_df_cleaned.dropna()

nb_df_cleaned.info()
location_df_cleaned.info()


nb_df_cleaned.to_csv('../data/processed/nb_cleaned.csv')
location_df_cleaned.to_csv('../data/processed/location_cleaned.csv')

nb_df_cleaned = pd.read_csv('../data/processed/nb_cleaned.csv')
location_df_cleaned = pd.read_csv('../data/processed/location_cleaned.csv')

location_df_cleaned = location_df_cleaned.head(50)
accepted_orders = location_df_cleaned[location_df_cleaned['driver_action'] == 'accepted']
unfulfilled_orders = location_df_cleaned[location_df_cleaned['driver_action'] == 'rejected']

def compute_geodesic_distance(coord1, coord2):
    return geodesic(coord1, coord2).meters

def count_within_radius(df, radius=500):
    count_data = []
    for i, row1 in df.iterrows():
        center_coord = (row1['lat'], row1['lng'])
        count_riders = 0
        count_orders = 0
        riders = set()
        orders = set()
        
        for j, row2 in df.iterrows():
            if i != j:
                coord = (row2['lat'], row2['lng'])
                distance = compute_geodesic_distance(center_coord, coord)
                if distance <= radius:
                    riders.add(row2['driver_id'])
                    orders.add(row2['order_id'])
        
        count_riders = len(riders)
        count_orders = len(orders)
        
        count_data.append({
            'center_order_id': row1['order_id'],
            'center_driver_id': row1['driver_id'],
            'count_riders_within_500m': count_riders,
            'count_orders_within_500m': count_orders
        })
    
    return pd.DataFrame(count_data)

accepted_counts = count_within_radius(accepted_orders)
unfulfilled_counts = count_within_radius(unfulfilled_orders)
