==> notebooks/02_feature_engineering.ipynb <==
import pandas as pd
from meteostat import Point, Daily
from datetime import datetime, timedelta
import requests
import ssl
from geopy.distance import geodesic

ssl._create_default_https_context = ssl._create_unverified_context






completed_orders= pd.read_csv('../data/processed/nb_cleaned.csv')
delivery_requests = pd.read_csv('../data/processed/location_cleaned.csv')



def parse_coordinates(address):
    """
    Parse the address to extract latitude and longitude.
    :param address: String in the format "latitude, longitude"
    :return: Tuple containing (latitude, longitude)
    """
    # Clean up the address string
    address = address.strip()  # Remove leading and trailing whitespace
    lat, lng = map(float, address.split(','))  # Split and convert to float
    return lat, lng


# Apply the function to the Trip Origin and Trip Destination columns
completed_orders[['Origin Lat', 'Origin Lng']] = completed_orders['Trip Origin'].apply(parse_coordinates).apply(pd.Series)
completed_orders[['Destination Lat', 'Destination Lng']] = completed_orders['Trip Destination'].apply(parse_coordinates).apply(pd.Series)



def calculate_distance(row):
    origin = (row['Origin Lat'], row['Origin Lng'])
    destination = (row['Destination Lat'], row['Destination Lng'])
    return geodesic(origin, destination).kilometers

completed_orders['trip_distance'] = completed_orders.apply(calculate_distance, axis=1)

print(completed_orders.head())


completed_orders.to_csv('../data/processed/nb_cleaned.csv', index=False)


Calculating the distance between th driver location and the trip origin for completed trips

completed_orders.info()


completed_orders['Trip Origin']
completed_orders[['Origin Lat', 'Origin Lng']] = completed_orders['Trip Origin'].apply(
    lambda x: pd.Series(map(float, x.split(',')))
)


completed_orders.head()

completed_orders = pd.read_csv('../data/processed/nb_cleaned.csv')
merged_df = delivery_requests.merge(
    completed_orders[['Trip ID', 'Origin Lat']], 
    left_on='order_id', 
    right_on='Trip ID', 
    how='left'
)


# columns_to_drop = ['Unnamed: 0', 'Trip Origin_x', 'origin_lat', 'origin_lng', 'Trip ID_y', 'Trip Origin_y', 'Trip ID_x']
# cleaned_df = merged_df.drop(columns=columns_to_drop)
merged_df = merged_df.merge(
     completed_orders[['Trip ID', 'Origin Lng', 'trip_distance']], 
    left_on='order_id', 
    right_on='Trip ID', 
    how='left'
)


merged_df.head(2)
columns_to_drop = ['Trip ID_x']
merged_df = merged_df.drop(columns=columns_to_drop)

merged_df.head(2)
merged_df.to_csv('safe.csv')



print(merged_df.isna().sum())
merged_df = merged_df.dropna(subset=['lat', 'lng', 'Origin Lat', 'Origin Lng'])



def calculate_driver_to_order_distance(row):
    driver_location = (row['lat'], row['lng'])
    order_location = (row['Origin Lat'], row['Origin Lng'])
    return geodesic(driver_location, order_location).kilometers


merged_df['driver_to_order_distance'] = merged_df.apply(calculate_driver_to_order_distance, axis=1)


merged_df.head()
# merged_df.to_csv('safe.csv')


driver_location = (6.602207,3.270465 )
order_location = (	6.601042, 3.276634)

# Calculate the distance in meters
distance = geodesic(driver_location, order_location).meters
print(distance)


delivery_requests.to_csv('../data/external/delivery_requests_with_distances.csv', index=False)


delivery_requests = pd.read_csv('delivery_requests_with_distances.csv')
delivery_requests.head(3)

delivery_requests.info()


# completed_orders.info()
merged_df.info()

completed_orders.info()

completed_orders.rename(columns={'Trip ID': 'order_id'}, inplace=True)
merged_df = pd.merge(merged_df, completed_orders[['order_id', 'Trip Start Time', 'Trip End Time']], on='order_id', how='left')
print(merged_df.head())


merged_df.loc[merged_df['driver_action'] == 'rejected', ['Trip Start Time', 'Trip End Time', 'trip_distance']] = 0
# merged_df.rename(columns={'Trip Start Time_x': 'Trip Start Time'}, inplace=True)
# merged_df.rename(columns={'Trip End Time_x': 'Trip End Time'}, inplace=True)
# columns_to_drop = [ 'Trip End Time_y', 'Trip Start Time_y']
# merged_df = merged_df.drop(columns=columns_to_drop)

merged_df.head(2)

merged_df['Trip Start Time'] = pd.to_datetime(merged_df['Trip Start Time'])
merged_df['Trip End Time'] = pd.to_datetime(merged_df['Trip End Time'])


merged_df['time_taken'] = (merged_df['Trip End Time'] - merged_df['Trip Start Time']).dt.total_seconds()
merged_df['speed'] = merged_df['trip_distance'] / merged_df['time_taken']
merged_df.head(2)

merged_df.to_csv('safe.csv')

Holidays and weekdays

import holidays
from datetime import date
merged_df = pd.read_csv('safe.csv')
merged_df.head(2)


merged_df['Trip End Time'] = pd.to_datetime(merged_df['Trip End Time']).dt.date 

# ng_holidays = holidays.Nigeria()

# def is_holiday_ng(trip_end_time):

#     return trip_end_time in ng_holidays

# def is_holiday_ng(trip_end_time):
#     return trip_end_time in ng_holidays

# merged_df['is_holiday'] = merged_df['Trip End Time'].apply(is_holiday_ng)


# merged_df = merged_df[merged_df['is_holiday'] == True]
# merged_df.head(2)

def classify_weekday_or_weekend(dt):
    if dt.weekday() < 5:  # Monday = 0, Sunday = 6
        return 'Weekday'
    else:
        return 'Weekend'

merged_df['weekday_or_weekend'] = merged_df['Trip End Time'].apply(classify_weekday_or_weekend)

print(merged_df[['Trip End Time', 'weekday_or_weekend']])


merged_df.to_csv('safe.csv')


from meteostat import Stations, Daily


merged_df['Trip Start Time'] = pd.to_datetime(merged_df['Trip Start Time']).dt.date 

def fetch_weather_data(lat, lon, start_date):
    # Find nearest weather station using Meteostat
    stations = Stations()
    station = stations.nearby(lat, lon).fetch(1)
    
    # Fetch daily weather data for the specified date
    data = Daily(station, start_date)
    data = data.fetch()
    
    return data

precipitation_data = []
for index, row in merged_df.iterrows():
    lat = row['Origin Lat']
    lon = row['Origin Lng']
    trip_start_time = row['Trip Start Time']
    
    # Fetch precipitation data for the specific date of the trip
    weather_data = fetch_weather_data(lat, lon, trip_start_time)
    
    # Extract precipitation information if available
    if not weather_data.empty:
        precipitation = weather_data.iloc[0]['prcp']
    else:
        precipitation = None
    
    precipitation_data.append(precipitation)
