==> notebooks/03_causal_inference.ipynb <==
import pandas as pd
import numpy as np
from causalnex.structure import StructureModel



from causalnex.plots import plot_structure
from causalnex.network import BayesianNetwork

from sklearn.model_selection import train_test_split
from causalnex.plots import plot_structure, NODE_STYLE, EDGE_STYLE

import matplotlib.pyplot as plt
from causalnex.structure.notears import from_pandas
from sklearn.preprocessing import LabelEncoder, MinMaxScaler

df = pd.read_csv('safe.csv')
df.head(3)

df = df.drop(columns=['Unnamed: 0.1', 'Unnamed: 0', 'Trip Start Time', 'Trip End Time'], errors='ignore')

df.head(2)
# df =df.fillna(df.mean())
# df = df.replace([float('inf'), float('-inf')], 1e10)


label_encoder = LabelEncoder()
df['driver_action'] = label_encoder.fit_transform(df['driver_action'])
df['weekday_or_weekend'] = label_encoder.fit_transform(df['weekday_or_weekend'])

scaler = MinMaxScaler()
scaled_df = pd.DataFrame(scaler.fit_transform(df), columns=df.columns)
train_df, holdout_df = train_test_split(df, test_size=0.2, random_state=42)

scaled_train_df = pd.DataFrame(scaler.fit_transform(train_df), columns=train_df.columns)
scaled_holdout_df = pd.DataFrame(scaler.transform(holdout_df), columns=holdout_df.columns)


driver_action_relations =[
    ('weekday_or_weekend', 'driver_action'),
    ('driver_to_order_distance', 'driver_action'),
    ('trip_distance', 'driver_action'),
    ('driver_to_order_distance', 'trip_distance')

]

scaled_train_df.info()

scaled_holdout_df.info()

sm = StructureModel()

sm.add_edges_from(driver_action_relations)



viz = plot_structure(
    sm,
    # graph_attributes={"scale": "0.5"},
    all_node_attributes=NODE_STYLE.WEAK,
    all_edge_attributes=EDGE_STYLE.WEAK,
)

viz.show('b.html')




from causalnex.network import BayesianNetwork
from causalnex.inference import InferenceEngine


bn = BayesianNetwork(sm)


bn.nodes

scaled_df.head(3)


train_df, holdout_df = train_test_split(scaled_df, test_size=0.2, random_state=42)

# print(train_df.tail(3))
train_df.describe()



bn = bn.fit_node_states_and_cpds(train_df)


ie = InferenceEngine(bn.fit_node_states_and_cpds(scaled_df))


for i in range(0,1):
  ie.do_intervention('weekday_or_weekend',i)
  print(f"predicted_outcome_{i} = {ie.query()['driver_action']}")

  ie.reset_do("weekday_or_weekend")

for i in range(0,1):
  ie.do_intervention('trip_distance',i)
  print(f"predicted_outcome_{i} = {ie.query()['driver_action']}")

ie.reset_do("trip_distance")

for i in range(0,10):
  ie.do_intervention('driver_to_order_distance',i)
  print(f"predicted_outcome_{i} = {ie.query()['driver_action']}")

  ie.reset_do("driver_to_order_distance")
